refactor: export hasText from shared
